Canvi tipus variable

+++ cecocloud-module/cecocloud-module-ecom/cecocloud-module-ecom-back-ecommerce/src/main/java/es/limit/cecocloud/ecom/back/ecommerce/logic/api/dto/PaycometNotification.java
@@ -68,10 +68,10 @@ public class PaycometNotification {
 	private String Currency;
 	
 	@NotNull
-	private int Amount;
+	private double Amount;
 	
 	@NotNull
-	private int AmountEur;
+	private double AmountEur;
 	
 	@NotNull
 	@Size(max = 255)
